cache tests and frontend update

+++ api/src/main/java/tqs/airquality/model/AirQualityData.java
@@ -11,8 +11,8 @@ public class AirQualityData implements Serializable {
     private double so2;
     private double no2;
     private double co;
-    private int pm10;
-    private int pm25;
+    private double pm10;
+    private double pm25;
     private int pollen_level_tree;
     private int pollen_level_grass;
     private int pollen_level_weed;
